enhance ArticleController and ArticleServiceImpl with new mutation methods for updating and deleting articles, and improve status update query in ArticleRepository

+++ Backend/Knowledge-Sharing-Platform/src/main/java/io/github/SahanChamara/service/impl/ArticleServiceImpl.java
@@ -84,8 +84,9 @@ public class ArticleServiceImpl implements ArticleService {
     @Override
     @Transactional
     public Boolean deleteArticle(Long id) {
+        if(!articleRepository.existsById(id)) return Boolean.FALSE;
         articleRepository.deleteById(id);
-        return true;
+        return Boolean.TRUE;
     }
 
     @Override
